adds a staged Dockerfile
It will build a small (360MB) docker image from the current source
tree. If not available, then it will also build the bap-dev image.

We will latter add the latter to the hub. Though I'm not sure what is
faster, dowloading 6GB or building new. I hope we can strip the dev
version (a version with the necessary toolchaing) to something closer
at least to 1 GB.