updated find-symbol plugin
Find-symbol plugin rewriten and uses new bap cmdline
interface facilities. Now, in terms of `bap-main`
library, it contains a command and a pass.
The command 'find-symbol' is responsible for the
foribidden symbols look up, and the pass for finding
recursive, complex and non structural functions.

Also, forbidden symbols look up returns almost
immediately, since it doesn't involve project creation.
The downside of it is that check 'forbidden-symbol' is no
longer composable with 'defective-symbol'.