uses the specification from the Image module
Prviously, we were using the specification from the `bap-llvm`
library. This specification is not published and it was just a hack,
and there aren't any benefits from using it.

Now we use the specification from the Image module.

Also, we used to emit `0x0` address for the forbidden external
symbols. This doesn't make any sence, so we won't provide an address
for the externals until `plt` entries naming will be fixed on the bap
side.