prevents taint occlusion in the untrusted-argument test
After atoi was properly stubed the untrusted-argument started to
fail. A deeper investigation revealed that it wasn't a regression
as previously the test results were incorrectly interpreted.

The problem with the existing version of the test is that it was using
the same value "42" as the taint source. Since the taint-input rules
taint the string passed to atoi as well as the return value, we had
"42" marked as untrusted. And since we're untrusting an already
untrusted value we had a report that was connecting the first call to
atoi in the good function (since it is the place where the global
value "42" was first tainted) to the bad usage in the function bad.

To prevent this occlusion, we use different values in good and bad
functions.